fix: move module with no srtup() to cog.core
