+++ cog/core/singleton.py
